feat(env): rename env var LAUNCH_BACKEND to MEGATRON_LAUNCH_BACKEND

+++ examples/aibenchmark/run_aibenchmark.sh
@@ -16,34 +16,71 @@ MEGATRON_PATH=$( dirname $( dirname ${CURRENT_DIR}))
 DTK_ENV=""                                                               # where env.sh of dtk
 DATA_PATH=""                                                             # path to redpajama_text_document
 TOKENIZER_MODEL_PATH=""                                                  # path to tokenizer.model
+LAUNCHER="mpirun"                                                        # mpirun or torchrun
 CHECKPOINT_PATH=""                                                       # path to ckpt
 NCCL_ENV=${MEGATRON_PATH}/requirements/env.sh                            # Please adjust the variables based on the actual NET being used
 LAUNCH_WITH_BINDING=${MEGATRON_PATH}/requirements/launch_with_binding.sh # Please adjust the variables based on the actual NET being used
 
-HOSTFILE="${hostfile_input}_slots"
-rm -f ${HOSTFILE} 
-cat ${hostfile_input} | sed -n "1,${node_num}p"|sed 's/$/ slots=8/' > ${HOSTFILE}
+if [[ "${LAUNCHER}" != "mpirun" && "${LAUNCHER}" != "torchrun" ]]; then
+    echo "Only mpirun and torchrun are supported as launch methods"
+    exit 1
+fi
 
-GPUS=$(($(cat ${HOSTFILE}|sort|uniq |wc -l)*8))
-HOST="$(cat ${HOSTFILE} |sed -n "1p"|awk -F ' ' '{print $1}')"
-PORT="25905"
+HOSTFILE="${hostfile_input}_slots"
+rm -f ${HOSTFILE}
+if [[ "${LAUNCHER}" == "mpirun" ]]; then
+    sed -n "1,${node_num}p" "${hostfile_input}" | sed 's/$/ slots=8/' > "${HOSTFILE}"
+else
+    sed -n "1,${node_num}p" "${hostfile_input}" | sed 's/$/ slots=1/' > "${HOSTFILE}"
+fi
+
+HOST="$(sed -n "1p" "${HOSTFILE}" | awk -F ' ' '{print $1}')"
+
+if [[ "${LAUNCHER}" == "mpirun" ]]; then
+    MPIRUN_NP=$((node_num * 8))
+    PORT=${PORT:-25905}
+else
+    MPIRUN_NP=${node_num}
+    GPUS_PER_NODE=${GPUS_PER_NODE:-8}
+    MASTER_ADDR=${MASTER_ADDR:-${HOST}}
+    MASTER_PORT=${MASTER_PORT:-25905}
+fi
+
+torchrun_args=""
+if [[ "${LAUNCHER}" == "torchrun" ]]; then
+    torchrun_args+="
+        -x PATH \
+        -x LD_LIBRARY_PATH \
+        -x PYTHONPATH \
+        -x MASTER_ADDR=${MASTER_ADDR} \
+        -x MASTER_PORT=${MASTER_PORT} \
+        -x NNODES=${node_num} \
+        -x GPUS_PER_NODE=${GPUS_PER_NODE} \
+    "
+fi
 
 # Runs aibenchmark model
 source ${NCCL_ENV}
-mpirun -np ${GPUS}  --hostfile ${HOSTFILE} \
-                    --allow-run-as-root \
-                    --bind-to none \
-                    --mca plm_rsh_no_tree_spawn 1 \
-                    bash -c "
-                    source ${DTK_ENV} && \
-                    source ${NCCL_ENV} && \
-                    ./train_aibenchmark_$((${GPUS} / 8))nodes.sh \
-                    ${HOST} \
-                    ${PORT} \
-                    --data_path=$DATA_PATH \
-                    --tokenizer_path=$TOKENIZER_MODEL_PATH \
-                    --checkpoint_path=$CHECKPOINT_PATH \
-                    --launch_with_binding=${LAUNCH_WITH_BINDING} \
-                    --profiling=$profiling" 2>&1 | tee log-$((${GPUS} / 8))nodes-`date +%F-%H%M`.log
 
+CMD="mpirun -np ${MPIRUN_NP} --hostfile ${HOSTFILE} \
+    --allow-run-as-root \
+    --bind-to none \
+    --mca plm_rsh_no_tree_spawn 1 \
+    ${torchrun_args} \
+    bash -c '
+    source ${DTK_ENV} && \
+    source ${NCCL_ENV} && \
+    cd ${CURRENT_DIR} && \
+    bash ${TRAIN_SCRIPT} \
+    ${HOST} \
+    ${PORT:-${MASTER_PORT}} \
+    --data_path=${DATA_PATH} \
+    --launch_backend=${LAUNCHER} \
+    --tokenizer_path=${TOKENIZER_MODEL_PATH} \
+    --checkpoint_path=${CHECKPOINT_PATH} \
+    --launch_with_binding=${LAUNCH_WITH_BINDING} \
+    --profiling=${profiling}' 2>&1 | tee log-${node_num}nodes-$(date +%F-%H%M).log
+"
+
+eval ${CMD}
 wait
